requirements.txt
asyncio>=3.4.3
fastapi>=0.115.8
httpx>=0.28.1
mypy>=1.15.0
pathlib>=1.0.1
pydantic-settings>=2.7.1
pytest>=8.3.4
trio>=0.29.0
uvicorn>=0.34.0